Виніс другу функцію розмітки з секції мейн для модульності
Вона знаходиться в тому самому файлі, що і функція розмітки 1 книги (BookMarkupLi)
