+++ src/js/main.js
@@ -1,7 +1,7 @@
 import Notiflix from 'notiflix';
 import { BooksAPI } from './books-api';
-import createBookMarkup from './bookMarkupLi';
-const booksList = document.querySelector('.bestsellers-list');
+import { createListMarkup } from './bookMarkupLi';
+
 const booksApi = new BooksAPI();
 async function onPageLoad() {
   try {
@@ -12,22 +12,3 @@ async function onPageLoad() {
   }
 }
 onPageLoad();
-function createListMarkup(books) {
-  const lists = [];
-  for (let i = 0; i < books.length; i += 1) {
-    const liMarkup = books[i].books
-      .map((book, idx) => {
-        if (idx < 5) {
-          return createBookMarkup(book);
-        }
-      })
-      .join('');
-    const markup = `<ul class="top-category-list">
-    <h2 class="top-category-title">${books[i].list_name}</h2>
-    <div class="top-books-wrapper">${liMarkup}</div>
-  </ul>
-  <button class="book-card-btn" type="button">See more</button>`;
-    lists.push(markup);
-  }
-  booksList.innerHTML = lists.join('');
-}
